Support XZ compressed SPDX Json Cache
Call scancode to scan source requires mess of CPU amd memory, in order
to save them, we use pre-built Json file as cache. And place
SPDX cache file into download layer which caused the size of download
layer was big.

In order to save the size of SPDX cache file, compress it as XZ

1. For scancode-to-spdx.py, support --output-spdx-json xxx.json.xz

2. For update-spdx.py, support --spdx-json-cache xxx.json.xz

3. Set SOURCE_SPDX_JSON_CACHE and PACKAGE_SPDX_JSON_CACHE as xz compressed

Issue: LINUXEXEC-33063

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>